pyproject.toml
[project]
name = "1c-zabbix-monitor_Windows_Linux"
version = "0.1.0"
description = "Универсальный агент на Python для мониторинга кластера 1С:Предприятие (8.3+) в Zabbix. Решение спроектировано с учетом высокой нагрузки: минимизирует количество обращений к кластеру и оптимизирует чтение логов."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "pydantic>=2.6.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.0.0",
]

[project.scripts]
"1c-zabbix-monitor" = "1c_zabbix_monitor_windows_linux_cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pip-audit>=2.7.0",
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pylint-pydantic>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/1c-zabbix-monitor_Windows_Linux"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pylint."MESSAGES CONTROL"]
disable = ["C0114", "C0115", "C0116", "R0903", "W0612"]

[tool.pylint.MASTER]
load-plugins = ["pylint_pydantic"]
